[2077] v16 Support boxing/unboxing operations (#2097)
* [2077] Support boxing/unboxing operations
[2065] Boxed enum types are treated as regular numbers
[1357] Boxed value types and String(s) should not be equal
[1292] Char (s) give incorrect types when stored as objects.
[1248] Doubles marked as objects don't give good types
[1301] Ushort (s), short (s), uint (s), byte (s) give incorrect types
[1290] Chars don't give correct ToString when stored as object

* [2077] Update Compiler Tests reference with box/unbox

* [2077] Update Portarell's Batch4 cleint tests Char, String and UInt32 props, [1579] [1600] [1611] appears to be fixed as well

* [2077] hasValue$1 function to be consistent with hasValue

* [2077] Client tests review/addition/verification

* [2077] Client tests review/addition/verification

* [2077] Xml doc for new [Unbox] attribute

* [2077] Compiler tests for [Unbox] attribute

+++ Tests/Batch3/BridgeIssues/1100/N1184.cs
@@ -34,10 +34,10 @@ namespace Bridge.ClientTest.Batch3.BridgeIssues
             Assert.AreEqual(typeof(float), f.GetType());
 
             object o = b;
-            Assert.AreEqual(typeof(int), o.GetType());
+            Assert.AreEqual(typeof(byte), o.GetType());
 
             o = f;
-            Assert.AreEqual(typeof(double), o.GetType());
+            Assert.AreEqual(typeof(float), o.GetType());
         }
     }
 }
